refactor: better names to schema files
