+++ spras/config/schema.py
@@ -7,7 +7,7 @@ from typing import Annotated, Optional
 
 from pydantic import AfterValidator, BaseModel, ConfigDict, Field
 
-from spras.config.raw_config_analysis import Analysis
+from spras.config.analysis_schema import Analysis
 from spras.config.util_enum import CaseInsensitiveEnum
 
 # The default length of the truncated hash used to identify parameter combinations
